requirements.txt
Flask==1.1.2
tensorflow-cpu
scikit-learn==0.23.2
gunicorn==20.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.19.5
pandas==1.2.4
requests==2.25.1